fix(callView): rename to .ts
