+++ src/App.vue
@@ -49,7 +49,7 @@ import Router from './router/router.js'
 import BrowserStorage from './services/BrowserStorage.js'
 import { EventBus } from './services/EventBus.ts'
 import { leaveConversationSync } from './services/participantsService.js'
-import { useCallViewStore } from './stores/callView.js'
+import { useCallViewStore } from './stores/callView.ts'
 import { useFederationStore } from './stores/federation.ts'
 import { useSidebarStore } from './stores/sidebar.ts'
 import { checkBrowser } from './utils/browserCheck.ts'
