+++ src/composables/useIsInCall.js
@@ -10,7 +10,7 @@ import { computed, onBeforeMount, onBeforeUnmount, ref } from 'vue'
 import { useStore } from './useStore.js'
 import { EventBus } from '../services/EventBus.ts'
 import SessionStorage from '../services/SessionStorage.js'
-import { useCallViewStore } from '../stores/callView.js'
+import { useCallViewStore } from '../stores/callView.ts'
 
 /**
  * Check whether the user joined the call of the current token in this PHP session or not
