+++ src/composables/useViewer.js
@@ -6,7 +6,7 @@
 import { nextTick, ref } from 'vue'
 
 import { useIsInCall } from './useIsInCall.js'
-import { useCallViewStore } from '../stores/callView.js'
+import { useCallViewStore } from '../stores/callView.ts'
 import { useSidebarStore } from '../stores/sidebar.ts'
 
 /**
